pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cast-studio"
version = "0.0.1"
description = "Convert asciinema .cast files to GIF and MP4, with a generic demo-recording engine"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Avi Zaguri" },]
keywords = ["asciinema", "cast", "gif", "mp4", "terminal", "demo", "recording"]

classifiers = [
#    "Development Status :: 5 - Production/Stable",
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Multimedia :: Video :: Conversion",
    "Topic :: Software Development :: Documentation",
]

dependencies = [
    "custom-python-logger>=2.0.13",
    "pillow>=12.1.1",
    "pre-commit>=4.5.0",
    "pytest>=9.0.1",
    "pytest-plugins>=2.0.0",
    "python-base-command>=0.1.3",
    "python-base-toolkit>=0.0.17",
    "setuptools>=80.9.0",
    "tenacity>=9.1.2",
    "wheel>=0.45.1",
]

[project.urls]
Homepage = "https://github.com/aviz92/cast-studio"
Repository = "https://github.com/aviz92/cast-studio"
Issues = "https://github.com/aviz92/cast-studio/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["cast_studio*"]
namespaces = false

[project.scripts]
cast        = "cast_studio.cli:main"
cast-render = "cast_studio.cli_render:main"
cast-init   = "cast_studio.cli_init:main"
